Trait pallet_mt_rpc::MerkleTreeRpcApiServer

pub trait MerkleTreeRpcApiServer<BlockHash, Element>: Sized + Send + Sync + 'static {
    fn get_leaves(
        &self, 
        tree_id: u32, 
        from: usize, 
        to: usize, 
        at: Option<BlockHash>
    ) -> RpcResult<Vec<Element>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        Element: Send + Sync + 'static + Serialize,
    { ... }
}

Server trait implementation for the MerkleTreeRpcApi RPC API.

Required Methods

fn get_leaves(
&self,
tree_id: u32,
from: usize,
to: usize,
at: Option<BlockHash>
) -> RpcResult<Vec<Element>>

Get The MerkleTree leaves.

This method calls into a runtime with Merkle pallet included and attempts to get the merkletree leaves. Optionally, a block hash at which the runtime should be queried can be specified.

Returns the (full) a Vec<[u8; 32]> of the leaves.

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
Element: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.